requirements.txt
# Core dependencies
discord.py>=2.3.2
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async file handling
aiofiles>=23.0.0

# Logging
structlog>=23.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0 